+++ First Crush/AppDelegate.swift
@@ -9,6 +9,7 @@
 import UIKit
 import AVFoundation
 import OneSignal
+import MediaPlayer;
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -114,6 +115,46 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
     }
 
+    override func remoteControlReceived(with event: UIEvent?)
+    {       print("Remote Event Received")
+        switch (event?.subtype) {
+        case UIEventSubtype.remoteControlTogglePlayPause?:
+            print("Received Headphone Play Pause")
+            MPRemoteCommandCenter.shared().togglePlayPauseCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
+                return .success
+            }
+            break;
+        case UIEventSubtype.remoteControlPlay?:
+            print("Received Remote Play")
+            MPRemoteCommandCenter.shared().playCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
+                return .success
+            }
+            break;
+        case UIEventSubtype.remoteControlPause?:
+            print("Received Remote Pause")
+            MPRemoteCommandCenter.shared().pauseCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
+                return .success
+            }
+            break;
+        case UIEventSubtype.remoteControlNextTrack?:
+            //Handle It
+            print("Received Next Event")
+            MPRemoteCommandCenter.shared().nextTrackCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
+                return .success
+            }
+            break;
+        case UIEventSubtype.remoteControlPreviousTrack?:
+            //Handle It
+            print("Received Previous Event")
+            MPRemoteCommandCenter.shared().previousTrackCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
+                return .success
+            }
+            break;
+        default:
+            break;
+        }
+    }
+    
 
 }
